API BREAK: use 'size_t' for counting bpline points
This pays down some compiler warning and cast debt, as well as moving closer to
uniformly using `size_t` for counting.

+++ lib/gvc/gvcjob.h
@@ -21,6 +21,7 @@ extern "C" {
 #include "gvcommon.h"
 #include "color.h"
 #include <stdbool.h>
+#include <stddef.h>
 
 #define ARRAY_SIZE(A) (sizeof(A)/sizeof(A[0]))
 
@@ -246,9 +247,9 @@ extern "C" {
 	pointf *url_map_p;
 
 	/* additional mapped regions for edges */
-	int url_bsplinemap_poly_n;      /* number of polygons in url bspline map
+	size_t url_bsplinemap_poly_n;      /* number of polygons in url bspline map
 					 if GVRENDER_DOES_MAPS && GVRENDER_DOES_MAP_BSPLINES */
-	int *url_bsplinemap_n;          /* array of url_bsplinemap_poly_n ints 
+	size_t *url_bsplinemap_n;          /* array of url_bsplinemap_poly_n ints 
 					 of number of points in each polygon */
 	pointf *url_bsplinemap_p;       /* all the polygon points */
